requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
anyio==4.3.0
apache-airflow==2.8.2
apache-airflow-providers-common-io==1.3.0
apache-airflow-providers-common-sql==1.11.0
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-http==4.9.1
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-sqlite==3.7.1
apispec==6.5.0
argcomplete==3.2.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
boto3==1.24.2
botocore==1.27.96
cachelib==0.9.0
cachetools==5.5.0
cassandra-driver==3.29.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
clickclick==20.10.2
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
colorlog==4.8.0
comm @ file:///C:/b/abs_67a8058udb/croot/comm_1709322909844/work
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.2
cryptography==42.0.5
debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
docutils==0.20.1
email-validator==1.3.1
exceptiongroup @ file:///C:/b/abs_c5h1o1_b5b/croot/exceptiongroup_1706031441653/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema==2.19.1
Flask==2.2.5
Flask-AppBuilder==4.3.11
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
geomet==0.2.1.post1
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-pubsub==2.23.0
google-re2==1.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.65.5
grpcio-status==1.65.5
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
img2table==1.3.0
importlib-metadata==6.11.0
importlib_resources==6.1.3
inflection==0.5.1
ipykernel @ file:///D:/bld/ipykernel_1708996677248/work
ipython @ file:///C:/b/abs_e5729i179y/croot/ipython_1694181400005/work
ipython-genutils==0.2.0
ipython-sql==0.5.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi @ file:///C:/ci/jedi_1644315428289/work
jellyfish==1.1.0
Jinja2==3.1.3
jmespath==1.0.1
json5==0.9.22
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_client @ file:///C:/b/abs_a6h3c8hfdq/croot/jupyter_client_1699455939372/work
jupyter_core @ file:///C:/b/abs_c769pbqg9b/croot/jupyter_core_1698937367513/work
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
lazy-object-proxy==1.10.0
limits==3.9.0
linkify-it-py==2.0.3
llvmlite==0.43.0
lockfile==0.12.2
lxml==5.3.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661915841596/work
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
more-itertools==10.2.0
multidict==6.0.5
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio @ file:///C:/b/abs_65d6lblmoi/croot/nest-asyncio_1708532721305/work
notebook==7.1.1
notebook_shim==0.2.4
numba==0.60.0
numexpr @ file:///C:/b/abs_5fucrty5dc/croot/numexpr_1696515448831/work
numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp39-cp39-win_amd64.whl#sha256=f284633067fde4bb078a09085a1c9e99a3218c129f5312b7b3d76c9340442a25
opencv-contrib-python==4.10.0.84
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
ordered-set==4.1.0
overrides==7.7.0
packaging @ file:///C:/b/abs_28t5mcoltc/croot/packaging_1693575224052/work
pandas==1.4.2
pandocfilters==1.5.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.4
pendulum==3.0.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pillow==11.0.0
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
playcrypt @ git+https://github.com/huseyingokay/playcrypt.git@79f0c8496cc75ab12d42a554017f7466d7dcfb3f
pluggy==1.4.0
praw==7.7.1
prawcore==2.4.0
prettytable==3.10.0
prison==0.2.1
prometheus_client==0.20.0
prompt-toolkit @ file:///C:/b/abs_68uwr58ed1/croot/prompt-toolkit_1704404394082/work
proto-plus==1.24.0
protobuf==5.27.3
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo @ file:///C:/b/abs_9ej7u6shci/croot/py-cpuinfo_1698068121579/work
pyarrow==8.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pycryptodome==3.17
Pygments @ file:///C:/b/abs_fay9dpq4n_/croot/pygments_1684279990574/work
PyJWT==2.8.0
pymongo==4.10.1
pypdfium2==4.30.0
python-daemon==3.0.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==2.0.7
python-nvd3==0.15.0
python-slugify==8.0.4
pytz==2024.1
pywin32==305.1
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq @ file:///C:/b/abs_89aq69t0up/croot/pyzmq_1705605705281/work
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rich-argparse==1.4.0
rpds-py==0.18.0
rsa==4.9
s3fs==0.4.0
s3transfer==0.6.2
Send2Trash==1.8.2
setproctitle==1.3.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tables @ file:///C:/b/abs_411740ajo7/croot/pytables_1705614883108/work
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.0
text-unidecode==1.3
time-machine==2.14.0
tinycss2==1.2.1
tomli==2.0.1
tornado @ file:///C:/b/abs_0cbrstidzg/croot/tornado_1696937003724/work
traitlets @ file:///C:/b/abs_e5m_xjjl94/croot/traitlets_1671143896266/work
trp==0.1.6a1
types-python-dateutil==2.8.19.20240106
typing_extensions @ file:///C:/b/abs_72cdotwc_6/croot/typing_extensions_1705599364138/work
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal-pathlib==0.1.4
update-checker==0.18.0
uri-template==1.3.0
urllib3==1.26.18
us==3.2.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.2.3
widgetsnbextension==4.0.10
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp @ file:///C:/b/abs_b0beoc27oa/croot/zipp_1704206963359/work
